NRF_FUNCS: Address Width getter and setter refactor.
Rename setPipeAddressWidth and getPipeAddressWidth to setAddressWidth
and getAddressWidth.
Refactor getAddressWidth to have only one return path using the
addr_witdh variable, this might be moved to be a global variable to
catch the address width of the nRF24 data pipes and check against it on
setters of the data pipes addresses.